+++ Gravity/src/com/evanreidland/e/client/graphics/GLRenderList.java
@@ -12,20 +12,20 @@ public class GLRenderList extends RenderList {
 		GL11.glBegin(GL11.GL_TRIANGLES);
 	}
 	public void addTri(Vertex a, Vertex b, Vertex c) {
-		GL11.glColor3f(a.r, a.g, a.b);
-		GL11.glNormal3f(a.normal.x, a.normal.y, a.normal.z);
-		GL11.glTexCoord2f(a.tx, a.ty);
-		GL11.glVertex3f(a.pos.x, a.pos.y, a.pos.z);
+		GL11.glColor3f((float)a.r, (float)a.g, (float)a.b);
+		GL11.glNormal3f((float)a.normal.x, (float)a.normal.y, (float)a.normal.z);
+		GL11.glTexCoord2f((float)a.tx, (float)a.ty);
+		GL11.glVertex3f((float)a.pos.x, (float)a.pos.y, (float)a.pos.z);
 		
-		GL11.glColor3f(b.r, b.g, b.b);
-		GL11.glNormal3f(b.normal.x, b.normal.y, b.normal.z);
-		GL11.glTexCoord2f(b.tx, b.ty);
-		GL11.glVertex3f(b.pos.x, b.pos.y, b.pos.z);
+		GL11.glColor3f((float)b.r, (float)b.g, (float)b.b);
+		GL11.glNormal3f((float)b.normal.x, (float)b.normal.y, (float)b.normal.z);
+		GL11.glTexCoord2f((float)b.tx, (float)b.ty);
+		GL11.glVertex3f((float)b.pos.x, (float)b.pos.y, (float)b.pos.z);
 		
-		GL11.glColor3f(c.r, c.g, c.b);
-		GL11.glNormal3f(c.normal.x, c.normal.y, c.normal.z);
-		GL11.glTexCoord2f(c.tx, c.ty);
-		GL11.glVertex3f(c.pos.x, c.pos.y, c.pos.z);
+		GL11.glColor3f((float)c.r, (float)c.g, (float)c.b);
+		GL11.glNormal3f((float)c.normal.x, (float)c.normal.y,(float) c.normal.z);
+		GL11.glTexCoord2f((float)c.tx, (float)c.ty);
+		GL11.glVertex3f((float)c.pos.x, (float)c.pos.y, (float)c.pos.z);
 	}
 	public void End() {
 		GL11.glEnd();
